Organização dos arquivos em pastas mais bem detalhadas.
