+++ backend/package.json
@@ -13,6 +13,6 @@
     "cors": "^2.8.5",
     "echarts-for-react": "^3.0.2",
     "express": "^4.18.2",
-    "mysql2": "^3.3.3"
+    "mysql2": "^3.6.0"
   }
 }
