+++ frontend/package.json
@@ -8,14 +8,14 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.4.0",
+    "axios": "^1.5.0",
     "bootstrap": "^5.3.0",
     "buffer": "^6.0.3",
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.11.2",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "^5.0.1",
     "react-toastify": "^9.1.3",
     "web-vitals": "^2.1.4",
