+++ frontend/src/App.js
@@ -2,20 +2,20 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import './App.css';
-import Login from './pages/Login';
-import Home from './pages/Home';
-import Produtos from './pages/Produtos';
-import EditarProduto from './pages/EditarProduto';
-import CadastroProduto from './pages/CadastroProduto';
-import ImportacaoNF from './pages/ImportacaoNF';
-import Usuarios from './pages/Usuarios';
-import EditarUsuario from './pages/EditarUsuario';
-import CadastroUsuario from './pages/CadastroUsuario';
-import RequisicoesRecusadas from './pages/RequisicoesRecusadas';
-import MonitoramentoRT from './pages/MonitoramentoRT';
-import Setor from './pages/Setor';
-import EditarSetor from './pages/EditarSetor';
-import CadastroSetor from './pages/CadastroSetor';
+import Login from './pages/Login.js';
+import Home from './pages/Home.js';
+import Produtos from './pages/Produtos.js';
+import EditarProduto from './pages/changes/EditarProduto.js';
+import CadastroProduto from './pages/inserts/CadastroProduto.js';
+import ImportacaoNF from './pages/ImportacaoNF.js';
+import Usuarios from './pages/Usuarios.js';
+import EditarUsuario from './pages/changes/EditarUsuario.js';
+import CadastroUsuario from './pages/inserts/CadastroUsuario.js';
+import RequisicoesRecusadas from './pages/RequisicoesRecusadas.js';
+import MonitoramentoRT from './pages/MonitoramentoRT.js';
+import Setor from './pages/Setor.js';
+import EditarSetor from './pages/changes/EditarSetor.js';
+import CadastroSetor from './pages/inserts/CadastroSetor.js';
 
 function App() {
   return (
