+++ frontend/src/componentes/SideBar.js
@@ -1,4 +1,4 @@
-import '../styles/SideBar.css';
+import './SideBar.css';
 import { Link } from 'react-router-dom';
 
 export default function SideBar() {
