+++ frontend/src/pages/StockDataGraphics.js
@@ -1,13 +1,6 @@
-// install (please try to align the version of installed @nivo packages)
-// yarn add @nivo/funnel
 import { ResponsiveFunnel } from '@nivo/funnel'
 
-// make sure parent container have a defined height when using
-// responsive component, otherwise height will be 0 and
-// no chart will be rendered.
-// website examples showcase many properties,
-// you'll often use just a few of them.
-export const MyResponsiveFunnel = ({ data /* see data tab */ }) => (
+export const StockDataGraphics = ({ data /* see data tab */ }) => (
     <ResponsiveFunnel
         data={data}
         margin={{ top: 20, right: 20, bottom: 20, left: 20 }}
