+++ frontend/src/pages/inserts/CadastroProduto.css
@@ -1,0 +1,37 @@
+.formulario-cadastrar{
+    margin-top: 100px;
+}
+
+.form-cadastro label{
+    display: block;
+}
+
+.form-cadastro input{
+    padding: 7px 10px;
+    border-radius: 5px;
+    border: none;
+    width: 200px;
+    background-color: rgb(226, 226, 226);
+    outline: none;
+}
+
+.form-cadastro div{
+    padding: 1vh 0;
+}
+
+.form-cadastro button{
+    margin: 1vh 0;
+    padding: 8px 40px;
+    width: 200px;
+    border: none;
+    outline: none;
+    background-color: #674CE7;
+    color: white;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: 0.3s ease-in-out;
+}
+
+.form-cadastro button:hover{
+    background-color: #a070ff;
+}
