--- frontend/src/styles/EditarProduto.css
@@ -1,37 +1,0 @@
-.formulario-editar{
-    margin-top: 100px;
-}
-
-.form-edit label{
-    display: block;
-}
-
-.form-edit input{
-    padding: 7px 10px;
-    border-radius: 5px;
-    border: none;
-    width: 200px;
-    background-color: rgb(226, 226, 226);
-    outline: none;
-}
-
-.form-edit div{
-    padding: 1vh 0;
-}
-
-.form-edit button{
-    margin: 1vh 0;
-    width: 200px;
-    padding: 8px 40px;
-    border: none;
-    outline: none;
-    background-color: #674CE7;
-    color: white;
-    border-radius: 5px;
-    cursor: pointer;
-    transition: 0.3s ease-in-out;
-}
-
-.form-edit button:hover{
-    background-color: #a070ff;
-}
