--- frontend/src/styles/Header.css
@@ -1,63 +1,0 @@
-*{
-    font-family: 'Poppins';
-    margin: 0 auto;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-.center{
-    margin: 0 auto;
-    padding: 0 2%;
-    text-align: center;
-}
-
-.clear{
-    clear: both;
-}
-
-.left{
-    float: left;
-}
-
-.right{
-    float: right;
-}
-
-.header-container{
-    position: fixed;
-    width: 100%;
-    height: 60px;
-    background: linear-gradient(to right, #6F00E6, #A55FF0);
-}
-
-.header-logo{
-    font-size: 26px;
-    color: white;
-    padding: 10px 15px;
-    transition: 0.3s ease-in-out;
-}
-
-.header-logo:hover{
-    transform: scale(1.03);
-}
-
-.center i{
-    cursor: pointer;
-    color: white;
-    font-size: 28px;
-    padding: 15px;
-    transition: 0.7s ease;
-}
-
-.center i:hover{
-    transform: scale(1.07);
-}
-
-i span{
-    font-weight: 200;
-    font-size: 10px;
-    display: block;
-    position: relative;
-    top: 2px;
-    left: 1px;
-}
