+++ frontend/src/styles/StockDataGraphics.css
